Fix time-conversion from ms to s on 32-bit systems

+++ controller/logcontroller.php
@@ -348,8 +348,8 @@ class LogController extends Controller {
 
                 // correct timestamp if needed
                 $time = $timestamp;
-                if (is_numeric($time) and (int)$time > 10000000000) {
-                    $time = (int)((int)$time / 1000);
+                if (is_numeric($time) and $time > 10000000000.0) {
+                    $time = (int)($time / 1000);
                 }
 
                 if (is_numeric($acc)) {
